[Test] rename dir
