+++ src/code/electron/nota.ts
@@ -12,7 +12,7 @@ import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/
 import { IpcServer } from "src/code/platform/ipc/browser/ipc";
 import { IpcChannel } from "src/code/platform/ipc/common/channel";
 import { ProxyChannel } from "src/code/platform/ipc/common/proxy";
-import { IMainLifecycleService, LifecyclePhase } from "src/code/platform/lifecycle1/electron/mainLifecycleService";
+import { IMainLifecycleService, LifecyclePhase } from "src/code/platform/lifecycle/electron/mainLifecycleService";
 import { StatusKey } from "src/code/platform/status/common/status";
 import { IMainStatusService } from "src/code/platform/status/electron/mainStatusService";
 import { IMainWindowService, MainWindowService } from "src/code/platform/window/electron/mainWindowService";
