+++ src/code/platform/configuration/common/abstractConfigService.ts
@@ -7,8 +7,8 @@ import { IConfigCollection } from "src/code/platform/configuration/common/config
 import { ConfigScope, IScopeConfigChangeEvent } from "src/code/platform/configuration/common/configRegistrant";
 import { IFileService } from "src/code/platform/files/common/fileService";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
-import { ILifecycleService } from "src/code/platform/lifecycle1/common/lifecycle";
-import { ILifecycleService as ILifecycleServiceDecorator } from "src/code/platform/lifecycle1/browser/browserLifecycleService";
+import { ILifecycleService } from "src/code/platform/lifecycle/common/lifecycle";
+import { ILifecycleService as ILifecycleServiceDecorator } from "src/code/platform/lifecycle/browser/browserLifecycleService";
 import { DataBuffer } from "src/base/common/file/buffer";
 
 export const IConfigService = createDecorator<IConfigService>('configuration-service');
