+++ src/code/platform/window/electron/mainWindowService.ts
@@ -7,7 +7,7 @@ import { IFileService } from "src/code/platform/files/common/fileService";
 import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 import { IInstantiationService } from "src/code/platform/instantiation/common/instantiation";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
-import { IMainLifecycleService } from "src/code/platform/lifecycle1/electron/mainLifecycleService";
+import { IMainLifecycleService } from "src/code/platform/lifecycle/electron/mainLifecycleService";
 import { ToOpenType, IUriToOpenConfiguration, IWindowConfiguration, IWindowCreationOptions } from "src/code/platform/window/common/window";
 import { IWindowInstance, WindowInstance } from "src/code/platform/window/electron/windowInstance";
 import { URI } from "src/base/common/file/uri";
